# feat: comprehensive PostgreSQL catalog and function support with performance optimization