--- a/src/catalog/constraint_populator.rs
+++ b/src/catalog/constraint_populator.rs
@@ -58,13 +58,16 @@ pub fn populate_constraints_for_table(conn: &Connection, table_name: &str) -> Re
     
     // Parse and populate constraints
     populate_table_constraints(conn, table_name, &create_sql, &table_oid)?;
-    
+
     // Parse and populate column defaults
     populate_column_defaults(conn, table_name, &create_sql, &table_oid)?;
-    
+
     // Populate indexes (including those created by UNIQUE constraints)
     populate_table_indexes(conn, table_name, &table_oid)?;
-    
+
+    // Populate dependencies (for Rails sequence ownership detection)
+    populate_table_dependencies(conn, table_name, &table_oid)?;
+
     info!("Successfully populated constraints for table: {}", table_name);
     Ok(())
 }
@@ -178,41 +181,94 @@ fn populate_column_defaults(conn: &Connection, table_name: &str, create_sql: &st
 
 /// Populate pg_index table with index information
 fn populate_table_indexes(conn: &Connection, table_name: &str, table_oid: &str) -> Result<()> {
-    // Get indexes for this table from sqlite_master
-    let mut stmt = conn.prepare("
-        SELECT name, sql FROM sqlite_master 
-        WHERE type = 'index' AND tbl_name = ?1 AND sql IS NOT NULL
-    ")?;
-    
-    let indexes = stmt.query_map([table_name], |row| {
-        Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
+    // First, get table columns to map column names to numbers (1-based like PostgreSQL)
+    let mut column_map = std::collections::HashMap::new();
+    let query = format!("PRAGMA table_info({})", table_name);
+    let mut column_stmt = conn.prepare(&query)?;
+    let column_rows = column_stmt.query_map([], |row| {
+        let cid: i32 = row.get(0)?;
+        let name: String = row.get(1)?;
+        Ok((name, cid + 1)) // Convert to 1-based like PostgreSQL attnum
     })?;
-    
-    for index_result in indexes {
-        let (index_name, index_sql) = index_result?;
+
+    for column_result in column_rows {
+        let (name, attnum) = column_result?;
+        column_map.insert(name, attnum);
+    }
+
+    // Get indexes using PRAGMA index_list
+    let query = format!("PRAGMA index_list({})", table_name);
+    let mut index_stmt = conn.prepare(&query)?;
+    let index_rows = index_stmt.query_map([], |row| {
+        let seq: i32 = row.get(0)?;
+        let name: String = row.get(1)?;
+        let is_unique: bool = row.get(2)?;
+        let origin: String = row.get(3)?;
+        let partial: bool = row.get(4)?;
+        Ok((seq, name, is_unique, origin, partial))
+    })?;
+
+    for index_result in index_rows {
+        let (_seq, index_name, is_unique, origin, _partial) = index_result?;
         let index_oid = generate_table_oid(&index_name);
-        
-        // Parse index information
-        let is_unique = index_sql.to_uppercase().contains("UNIQUE");
-        let is_primary = index_name.contains("primary") || index_name.contains("pkey");
-        
+
+        // Skip auto-indexes created by SQLite for unique constraints
+        if index_name.starts_with("sqlite_") {
+            continue;
+        }
+
+        // Get column information for this index using PRAGMA index_info
+        let query = format!("PRAGMA index_info({})", index_name);
+        let mut info_stmt = conn.prepare(&query)?;
+        let info_rows = info_stmt.query_map([], |row| {
+            let seqno: i32 = row.get(0)?;
+            let cid: i32 = row.get(1)?;
+            let name: Option<String> = row.get(2)?;
+            Ok((seqno, cid, name))
+        })?;
+
+        let mut column_numbers = Vec::new();
+        let mut column_count = 0;
+
+        for info_result in info_rows {
+            let (_seqno, _cid, col_name_opt) = info_result?;
+            if let Some(col_name) = col_name_opt
+                && let Some(&attnum) = column_map.get(&col_name) {
+                column_numbers.push(attnum.to_string());
+                column_count += 1;
+            }
+        }
+
+        // Build indkey field (space-separated column numbers, PostgreSQL format)
+        let indkey = column_numbers.join(" ");
+
+        // Determine if this is a primary key index
+        let is_primary = origin == "pk" || index_name.contains("primary") || index_name.contains("pkey");
+
         conn.execute(
             "INSERT OR IGNORE INTO pg_index (
-                indexrelid, indrelid, indnatts, indnkeyatts, 
-                indisunique, indisprimary
-            ) VALUES (?1, ?2, 1, 1, ?3, ?4)",
+                indexrelid, indrelid, indnatts, indnkeyatts,
+                indisunique, indisprimary, indkey,
+                indisexclusion, indimmediate, indisclustered,
+                indisvalid, indcheckxmin, indisready, indislive,
+                indisreplident, indcollation, indclass, indoption,
+                indexprs, indpred
+            ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, 0, 1, 0, 1, 0, 1, 1, 0, '', '', '', '', '')",
             [
                 &index_oid,
                 table_oid,
+                &column_count.to_string(),
+                &column_count.to_string(), // For regular indexes, indnkeyatts = indnatts
                 &(is_unique as i32).to_string(),
                 &(is_primary as i32).to_string(),
+                &indkey,
             ]
         )?;
-        
-        debug!("Inserted index: {} (unique: {}, primary: {}) for table: {}", 
-               index_name, is_unique, is_primary, table_name);
+
+        debug!("Inserted index: {} (unique: {}, primary: {}, columns: {}) for table: {}",
+               index_name, is_unique, is_primary, indkey, table_name);
     }
-    
+
     Ok(())
 }
 
@@ -440,4 +496,75 @@ fn get_column_number(create_sql: &str, target_column: &str) -> Option<i16> {
         }
     
     None
+}
+
+/// Populate pg_depend table with dependency information for sequence ownership detection
+fn populate_table_dependencies(conn: &Connection, table_name: &str, table_oid: &str) -> Result<()> {
+    debug!("Populating dependencies for table: {}", table_name);
+
+    // Get table columns to find INTEGER PRIMARY KEY (acts like SERIAL in SQLite)
+    let query = format!("PRAGMA table_info({})", table_name);
+    let mut column_stmt = conn.prepare(&query)?;
+    let column_rows = column_stmt.query_map([], |row| {
+        let cid: i32 = row.get(0)?;
+        let name: String = row.get(1)?;
+        let column_type: String = row.get(2)?;
+        let pk: i32 = row.get(5)?;
+        Ok((cid, name, column_type, pk))
+    })?;
+
+    let mut pk_columns = Vec::new();
+    for (cid, column_name, column_type, pk) in column_rows.flatten() {
+        debug!("Column: {} (cid={}, type={}, pk={}) in table {}", column_name, cid, column_type, pk, table_name);
+        if pk > 0 {
+            pk_columns.push((cid, column_name.clone(), column_type.clone(), pk));
+        }
+    }
+
+    // Only create dependencies for single-column INTEGER PRIMARY KEY
+    if pk_columns.len() == 1 {
+        let (cid, column_name, column_type, _pk) = &pk_columns[0];
+        if column_type.to_uppercase().contains("INTEGER") {
+            debug!("Found single INTEGER PRIMARY KEY column: {} in table {} at position {}", column_name, table_name, cid + 1);
+
+            // Generate deterministic OIDs
+            let sequence_oid = generate_sequence_oid(table_name, column_name);
+            let table_oid_str = table_oid; // table_oid is already a string
+
+            // Insert dependency record into pg_depend table
+            // This represents: sequence depends on table column (automatic dependency)
+            let result = conn.execute(
+                "INSERT OR REPLACE INTO pg_depend (classid, objid, objsubid, refclassid, refobjid, refobjsubid, deptype)
+                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
+                rusqlite::params![
+                    "1259",        // classid: pg_class OID (for sequences)
+                    sequence_oid.to_string(),  // objid: sequence OID
+                    "0",           // objsubid: 0 for sequences
+                    "1259",        // refclassid: pg_class OID (for tables)
+                    table_oid_str,             // refobjid: table OID
+                    cid + 1,       // refobjsubid: column number (1-based like PostgreSQL)
+                    "a"            // deptype: automatic dependency
+                ],
+            )?;
+
+            debug!("Inserted pg_depend record: sequence {} depends on column {} of table {} (result: {})",
+                   sequence_oid, column_name, table_name, result);
+        }
+    } else {
+        debug!("Table {} has {} PK columns, skipping dependency creation", table_name, pk_columns.len());
+    }
+
+    Ok(())
+}
+
+/// Generate a deterministic OID for a sequence based on table and column name
+fn generate_sequence_oid(table_name: &str, column_name: &str) -> u32 {
+    use std::collections::hash_map::DefaultHasher;
+    use std::hash::{Hash, Hasher};
+
+    let sequence_name = format!("{}_{}_seq", table_name, column_name);
+    let mut hasher = DefaultHasher::new();
+    sequence_name.hash(&mut hasher);
+    let hash = hasher.finish();
+    32768 + ((hash % 65536) as u32) // Different range from tables to avoid conflicts
 }