--- a/src/catalog/mod.rs
+++ b/src/catalog/mod.rs
@@ -2,6 +2,7 @@
 pub mod query_interceptor;
 pub mod pg_class;
 pub mod pg_attribute;
+pub mod pg_constraint;
 pub mod pg_enum;
 pub mod system_functions;
 pub mod where_evaluator;