--- a/src/catalog/mod.rs
+++ b/src/catalog/mod.rs
@@ -9,6 +9,7 @@ pub mod pg_proc;
 pub mod pg_description;
 pub mod pg_roles;
 pub mod pg_user;
+pub mod pg_stats;
 pub mod system_functions;
 pub mod where_evaluator;
 pub mod constraint_populator;