--- a/src/query/extended_fast_path.rs
+++ b/src/query/extended_fast_path.rs
@@ -22,6 +22,7 @@ impl ExtendedFastPath {
     }
     
     /// Execute a parameterized query using prepared statements directly
+    #[allow(clippy::too_many_arguments)]
     pub async fn execute_with_params<T>(
         framed: &mut Framed<T, crate::protocol::PostgresCodec>,
         db: &Arc<DbHandler>,