--- a/tests/datetime_working_test.rs
+++ b/tests/datetime_working_test.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::collapsible_if)]
+
 mod common;
 use common::setup_test_server;