… CLAUDE.md
… CONTRIBUTING.md
… Cargo.lock
… README.md
… TODO.md
… benches/simple_query_bench.rs
… benchmarks/README.md
… benchmarks/benchmark_array_binary.py
… benchmarks/run_array_benchmark.sh
… benchmarks/test_array_benchmark_simple.py
… src/catalog/pg_attribute.rs
… src/catalog/pg_depend.rs
… src/catalog/pg_description.rs
… src/catalog/pg_proc.rs
… src/catalog/pg_roles.rs
… src/catalog/pg_stats.rs
… src/catalog/pg_user.rs
… src/catalog/system_functions.rs
… src/functions/system_functions.rs
… src/lib.rs
… src/migration/registry.rs
… src/optimization/statement_cache_optimizer.rs
… src/protocol/binary.rs
… src/protocol/codec.rs
… src/protocol/value_handler.rs
… src/query/executor.rs
… src/query/extended.rs
… src/query/fast_path.rs
… src/query/lazy_processor.rs
… src/query/query_processor.rs